Rename alembic.py -> schema.py to remove name clash
It conflicts with the system package so the `alembic.*` imports inside
`alembic.py` wouldn't work.
