+++ scripts/schema.py
